requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
MouseInfo==0.1.3
packaging==24.2
pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
pywinusb==0.4.2
